[기능] 유저관리 탭을 반응형 디자인으로 변경
