+++ src/modules/users/UserPage.vue
@@ -1,17 +1,41 @@
 <script setup lang="ts">
-import UserDiffListAndRegisterTab from '@^users/components/UserDiffListAndRegisterTab.vue';
+import CommonUserPageOnDesktop from '@^users/components/CommonUserPageOnDesktop.vue';
 import UserListAndSearchBox from '@^users/components/UserListAndSearchBox.vue';
 </script>
 
 <template>
-  <section class="w-100 h-100 d-flex flex-row gap-3">
-    <section class="w-50 h-100">
-      <UserListAndSearchBox></UserListAndSearchBox>
-    </section>
-    <section class="w-50 h-100">
-      <UserDiffListAndRegisterTab></UserDiffListAndRegisterTab>
-    </section>
+  <section class="desktop-frame">
+    <CommonUserPageOnDesktop></CommonUserPageOnDesktop>
+  </section>
+  <section class="mobile-frame">
+    <UserListAndSearchBox></UserListAndSearchBox>
   </section>
 </template>
 
-<style lang="scss" scoped></style>
+<style lang="scss" scoped>
+@include media-breakpoint-between('mobile', 'tablet-landscape') {
+  .desktop-frame {
+    display: none;
+  }
+
+  .mobile-frame {
+    display: block;
+
+    width: 100%;
+    height: 100%;
+  }
+}
+
+@include media-breakpoint-up('tablet-landscape') {
+  .desktop-frame {
+    display: block;
+
+    width: 100%;
+    height: 100%;
+  }
+
+  .mobile-frame {
+    display: none;
+  }
+}
+</style>
