+++ src/modules/users/components/UserRegisterConfirmModal.vue
